Consolidate promotor controllers
